refactor(sparse): scope payload module to the expression layer, realize constraints via CSRConstraint.from_payload
